stalld:  first attempt at emulating DEADLINE using FIFO
This is a first cut at the logic to do stalled thread boosting using
SCHED_FIFO, when running on a kernel without SCHED_DEADLINE support.
The idea is that in the boost logic, we change the starving thread's
policy to SCHED_FIFO:98, then sleep up to the 'period' boundary (same
period used for DEADLINE), then change the thread back to it's original
policy and parameters.

Signed-off-by: Clark Williams <williams@redhat.com>