use FIFO for boosting (v2)
This is a set of changes that allow stalld to operation without the
Deadline scheduler policy. The idea is to emulate how deadline
works boosting a starving thread. When a boost operation is called
for, the fifo boosting logic figures out how many 'periods' would
be in the specified boost duration (default 3 seconds) and iterates
for that number of periods, boosting the starving thread to fifo:98,
sleeping for the 'runtime', deboosting to the original policy, sleeping
for the remainder (period - runtime) and repeating until the boost
duration is done.

The FIFO boost logic is turned on when the running kernel does not
support DEADLINE or when the command line option -F/--force_fifo is
specified.

Signed-off-by: Clark Williams <clark.williams@gmail.com>