bpf: Reduce arguments to log() to fix verifier error
On some kernel versions, the BPF verifier complains that the argument
list to the log() macro is too long. This prevents the BPF program from
loading.

Remove the scheduling class ("rt" or "fair") from the log_task_prefix()
macro. This reduces the number of arguments and allows the BPF program
to be verified and loaded successfully on affected kernels.d from the
log message, reducing the number of arguments to a valid count.

Signed-off-by: Wander Lairson Costa <wander@redhat.com>