stalld: Always reset starving vector in merge_taks_info
The merge_taks_info() function updates the starving task information
for a given CPU. It is supposed to clear the starving vector and then
re-populate it with the new information.

However, the vector was only being cleared if the number of new tasks
was zero. If there were new tasks, but none of them met the conditions
to be considered starving, the vector would not be cleared or updated,
causing it to retain stale data from the previous interval.

Unconditionally call the update_cpu_starving_vector() function,
ensuring the starving vector is always reset before the new state is
computed. This prevents stale data from persisting and leading to
incorrect throttling decisions.

Resolves: RHEL-114506

Signed-off-by: Wander Lairson Costa <wander@redhat.com>