stalld:  make the queue_track (eBPF) backend the default
Change the default backend to be the queue_track eBPF backend.

This is mainly for performance and scalability since the access
of the eBPF map containing run-queue information is much faster
than parsing the output of the /sys/kernel/debug/sched/debug
file.

The i686 and ppc64le arches don't support eBPF so default them
to using the sched_debug backend.

Signed-off-by: Clark Williams <williams@redhat.com>